Switch all Jamulus audio sample processing to use floats instead of a mix of double and int16_t .
MacOS, Linux and Android already do this, and ASIO also supports it.
Change recording format to be 24-bit WAV, to get the full precision
of the mixed audio!

This patch gives the user the full resolution of the audio device,
both when receiving audio and transmitting audio.

Cleanup LSB/MSB sample processing in ASIO driver while at it.
LSB and MSB indicate little-endian and big-endian data format
presumably.

+++ src/recorder/jamrecorder.h
@@ -71,7 +71,7 @@ class CJamClient : public QObject
 public:
     CJamClient(const qint64 frame, const int numChannels, const QString name, const CHostAddress address, const QDir recordBaseDir);
 
-    void Frame(const QString name, const CVector<int16_t>& pcm, int iServerFrameSizeSamples);
+    void Frame(const QString name, const CVector<float>& pcm, int iServerFrameSizeSamples);
 
     void Disconnect();
 
@@ -107,7 +107,7 @@ public:
 
     CJamSession(QDir recordBaseDir);
 
-    void Frame(const int iChID, const QString name, const CHostAddress address, const int numAudioChannels, const CVector<int16_t> data, int iServerFrameSizeSamples);
+    void Frame(const int iChID, const QString name, const CHostAddress address, const int numAudioChannels, const CVector<float> data, int iServerFrameSizeSamples);
 
     void End();
 
@@ -198,7 +198,7 @@ public slots:
     /**
      * @brief Handle a frame of data to process
      */
-    void OnFrame ( const int iChID, const QString name, const CHostAddress address, const int numAudioChannels, const CVector<int16_t> data );
+    void OnFrame ( const int iChID, const QString name, const CHostAddress address, const int numAudioChannels, const CVector<float> data );
 };
 
 }
